Fix postgres installation check
Fix the postgres installation check so it can pass.

The output of 'command -v postgres' when 'postgres' is installed via Homebrew is not reflected in the 
current regular expression passed to grep, so the test for postgres installation used to always fail. 
Remove the string '(PostgreSQL)' from the regex passed to grep so the test can pass.